# references/mfa/openapi.yaml
openapi: 3.0.3
info:
  title: MFA OTP API
  description: API for storing and retrieving MFA tokens. Built with Cloudflare Workers.
  version: 0.0.1
paths:
...
  /api/validate:
    post:
      summary: Validate OTP
      description: Validates an OTP token for login. Requires MFA to already be verified and enabled.
      operationId: validate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '200':
          description: OTP validated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/GuidNotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with guid claim
  schemas:
    OtpRequest:
      type: object
      required:
        - otp
      properties:
        otp:
          type: string
          description: 6-digit OTP code from authenticator app
...
    ValidateSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OTP validated.
...
  responses:
    Unauthorized:
      description: Invalid or missing token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    GuidNotFound:
      description: User GUID not found
    ServerError:
      description: Internal server error